pyproject.toml
[tool.poetry]
name = "treineeeeeee"
version = "0.1.0"
description = ""
authors = ["frutez <glebka.yatsenko@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
djangorestframework = "^3.14.0"
folium = "^0.15.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
